+++ CHANGELOG.md
@@ -18,6 +18,10 @@ At the moment this project **does not** adhere to
   cleaned up. A lot of storage entries, events, and extrinsics were removed from the `Registry`
   pallet. The genesis build config was also removed. Additionally, the `new/user/` HTTP endpoint in
   the TSS was removed since it was no longer necessary.
+- In [#1031](https://github.com/entropyxyz/entropy-core/pull/1031), more Staking calls were blocked
+  to go through the `staking_extention` pallet. This makes sure no funds can be unbonded from a
+  validator if they are currently in the signing comittee. This was applied to `unbond`, `chill`,
+  and `withdraw_unbonded`
 - In [#1045](https://github.com/entropyxyz/entropy-core/pull/1045), `ProgramsInfo` now takes `version_number` to maintain backwards compatibility if programs runtime is updated
 - In [#1050](https://github.com/entropyxyz/entropy-core/pull/1050), the flow for signing has changed.
   A user now sends their request to any validator that is not a signer. This will act as a relayer.
@@ -45,6 +49,7 @@ At the moment this project **does not** adhere to
 
 ### Changed
 - Fix TSS `AccountId` keys in chainspec ([#993](https://github.com/entropyxyz/entropy-core/pull/993))
+- No unbonding when signer or next signer ([#1031](https://github.com/entropyxyz/entropy-core/pull/1031))
 - Add relay tx endpoint ([#1050](https://github.com/entropyxyz/entropy-core/pull/1050))
 
 ### Removed